Register BlobStoreApi into BlobStoreCoreExtension and merge blob-core with blob-api
