+++ extensions/azure/blobstorage/blob-core/src/main/java/org/eclipse/dataspaceconnector/azure/blob/core/api/BlobStoreApiImpl.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.common.azure;
+package org.eclipse.dataspaceconnector.azure.blob.core.api;
 
 import com.azure.core.util.BinaryData;
 import com.azure.storage.blob.BlobServiceClient;
