+++ extensions/azure/blobstorage/blob-data-operator/build.gradle.kts
@@ -21,7 +21,6 @@ val storageBlobVersion: String by project
 dependencies {
     api(project(":spi"))
     api(project(":extensions:inline-data-transfer:inline-data-transfer-spi"))
-    api(project(":extensions:azure:blobstorage:blob-api"))
 
     implementation(project(":extensions:azure:blobstorage:blob-core"))
 }
