+++ extensions/azure/blobstorage/blob-data-operator/src/main/java/org/eclipse/dataspaceconnector/azure/blob/operator/BlobStoreReader.java
@@ -1,7 +1,7 @@
 package org.eclipse.dataspaceconnector.azure.blob.operator;
 
 import org.eclipse.dataspaceconnector.azure.blob.core.AzureBlobStoreSchema;
-import org.eclipse.dataspaceconnector.common.azure.BlobStoreApi;
+import org.eclipse.dataspaceconnector.azure.blob.core.api.BlobStoreApi;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.eclipse.dataspaceconnector.spi.types.domain.DataAddress;
 import org.eclipse.dataspaceconnector.transfer.inline.spi.DataReader;
