+++ extensions/azure/blobstorage/blob-provision/build.gradle.kts
@@ -22,7 +22,6 @@ val jodahFailsafeVersion: String by project
 
 dependencies {
     api(project(":spi"))
-    api(project(":extensions:azure:blobstorage:blob-api"))
     api(project(":extensions:azure:blobstorage:blob-core"))
 
     implementation("com.azure:azure-storage-blob:${storageBlobVersion}")
