+++ samples/05-file-transfer-cloud/transfer-file/src/main/java/org/eclipse/dataspaceconnector/extensions/transfer/CloudTransferExtension.java
@@ -3,9 +3,9 @@ package org.eclipse.dataspaceconnector.extensions.transfer;
 import net.jodah.failsafe.RetryPolicy;
 import org.eclipse.dataspaceconnector.aws.s3.operator.S3BucketReader;
 import org.eclipse.dataspaceconnector.aws.s3.operator.S3BucketWriter;
+import org.eclipse.dataspaceconnector.azure.blob.core.api.BlobStoreApi;
 import org.eclipse.dataspaceconnector.azure.blob.operator.BlobStoreReader;
 import org.eclipse.dataspaceconnector.azure.blob.operator.BlobStoreWriter;
-import org.eclipse.dataspaceconnector.common.azure.BlobStoreApi;
 import org.eclipse.dataspaceconnector.dataloading.AssetLoader;
 import org.eclipse.dataspaceconnector.spi.asset.DataAddressResolver;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
@@ -71,6 +71,4 @@ public class CloudTransferExtension implements ServiceExtension {
 
         assetIndex.accept(asset, dataAddress);
     }
-
-
 }
