+++ settings.gradle.kts
@@ -48,7 +48,6 @@ include(":extensions:aws:aws-test")
 include(":extensions:azure:blobstorage")
 include(":extensions:azure:blobstorage:blob-core")
 include(":extensions:azure:blobstorage:blob-provision")
-include(":extensions:azure:blobstorage:blob-api")
 include(":extensions:azure:blobstorage:blob-data-operator")
 include(":extensions:azure:events")
 include(":extensions:azure:events-config")
